// app/__generated__/CommentThread_issue.graphql.ts
export type CommentThread_issue$data = {
  readonly comments: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly nodeId: string;
        readonly " $fragmentSpreads": FragmentRefs<"CommentItem_comment">;
      };
    }>;
    readonly totalCount: number;
  } | null | undefined;
  readonly id: string;
  readonly nodeId: string;
  readonly " $fragmentType": "CommentThread_issue";
};
export type CommentThread_issue$key = {
  readonly " $data"?: CommentThread_issue$data;
  readonly " $fragmentSpreads": FragmentRefs<"CommentThread_issue">;
};